bug(feedback-implementation): rename files and remove unwaanted comments [Delivers #157977822]
