+++ server/controllers/admin.js
@@ -1,6 +1,6 @@
-import reqHelper from '../helpers/reqHelper';
+import reqHelper from '../helpers/requestHelper';
 import db from '../models/testData';
-import databaseLink from '../models/databaseLink';
+import databaseLink from '../models/databaseConnection';
 
 const dotenv = require('dotenv');
 
@@ -32,7 +32,7 @@ export default class adminController {
       success: false,
       message: 'Failed to retrieved request for user',
     });
-  }// getUserRequest end
+  }
 
   /**
  * API method GET all request by user
